vita-system(backend)/requirements.txt
annotated-types==0.6.0
anyio==3.7.1
blinker==1.6.3
click==8.1.7
colorama==0.4.6
ecdsa==0.18.0
exceptiongroup==1.1.3
fastapi==0.104.0
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
greenlet==3.0.1
h11==0.14.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mysqlclient==2.2.0
paho-mqtt==1.6.1
pyasn1==0.5.0
pydantic==2.4.2
pydantic-settings==2.0.3
pydantic_core==2.10.1
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.22
starlette==0.27.0
typing_extensions==4.8.0
uvicorn==0.23.2
Werkzeug==3.0.1
